# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shumi API",
        "title": "Shumi API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/namespace/{namespace_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Получить ленту из анкет пользователей неймспейса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id неймспейса",
                        "name": "namespace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.PageData"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Зайти в неймспейс",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id неймспейса",
                        "name": "namespace",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "инвайт код для входа в неймспейс",
                        "name": "invite_code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "403": {
                        "description": "Неверный код входа",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Выйти из неймспейса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id неймспейса",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
        "/api/tg": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Проверить существование пользователя в бд",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Проверить существование пользователя в бд",
                "parameters": [
                    {
                        "description": "tg id пользователя",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "tg username пользователя",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "tg firstname пользователя",
                        "name": "firstname",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Скрыть/показывать анкету другим пользователям",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Скрыть/показывать анкету другим пользователям",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.Error": {
            "description": "Ошибка",
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "database.User": {
            "description": "Основные данные профиля и связанные виджеты.",
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hide": {
                    "type": "boolean"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "miniapp_name": {
                    "type": "string"
                },
                "tg_user": {
                    "type": "string"
                },
                "user_info": {
                    "type": "string"
                },
                "user_photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Widget"
                    }
                }
            }
        },
        "database.Vote": {
            "description": "Голос в опросе",
            "type": "object",
            "properties": {
                "option": {
                    "type": "integer"
                },
                "survey": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "database.Widget": {
            "description": "Виджет",
            "type": "object",
            "properties": {
                "additionalData": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/database.User"
                },
                "vote": {
                    "$ref": "#/definitions/database.Vote"
                }
            }
        },
...
        "handlers.PageData": {
            "description": "Данные анкеты (все поля database.User + tg username)",
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hide": {
                    "type": "boolean"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "miniapp_name": {
                    "type": "string"
                },
                "tg_user": {
                    "type": "string"
                },
                "user_info": {
                    "type": "string"
                },
                "user_photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "username": {
                    "type": "string"
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Widget"
                    }
                }
            }
        }
    }
}